requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
bcolz-zipline==1.2.6
beautifulsoup4==4.12.3
blessed==1.20.0
boto3==1.34.79
botocore==1.34.79
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
croniter==1.3.15
cycler==0.12.1
deepdiff==6.7.1
docker-pycreds==0.4.0
editor==1.6.6
exceptiongroup==1.2.0
fastapi==0.110.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.8.4
h11==0.14.0
idna==3.6
inquirer==3.2.4
itsdangerous==2.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lightning==1.8.6
lightning-cloud==0.5.65
lightning-utilities==0.11.2
markdown-it-py==3.0.0
matplotlib==3.8.4
mdurl==0.1.2
multidict==6.0.5
mxnet-mkl==1.6.0
numpy==1.23.1
opencv-python==4.9.0.80
ordered-set==4.1.0
packaging==24.0
pandas==2.2.1
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==1.8.6
pytz==2024.1
PyYAML==6.0.1
readchar==4.0.6
requests==2.31.0
rich==13.7.1
runs==1.2.2
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.13.0
sentry-sdk==1.44.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
starsessions==1.3.0
tensorboardX==2.6.2.2
threadpoolctl==3.4.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchmetrics==1.3.2
torchvision==0.14.1+cu117
tqdm==4.66.2
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wandb==0.16.6
wcwidth==0.2.13
websocket-client==1.7.0
xmod==1.8.1
yarl==1.9.4
